Cleaned and renamed scripts
